refactored entity and defs code into headers
